test: rename docker test, shorten api call
